requirements.txt
torch>=2.1.0
torchvision>=0.16.0
einops>=0.7.0
numpy>=1.24.0
pillow>=10.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
tensorboard>=2.14.0
pyyaml>=6.0
tqdm>=4.65.0
medmnist
